+++ src/simulation_management_with_reliability/record_query_reliable_pdu.rs
@@ -58,7 +58,7 @@ impl Pdu for RecordQueryReliablePdu {
         self.originating_entity_id.serialize(buf);
         self.receiving_entity_id.serialize(buf);
         buf.put_u32(self.request_id);
-        buf.put_u8(self.required_reliability_service as u8);
+        buf.put_u8(self.required_reliability_service);
         buf.put_u16(self.pad1);
         buf.put_u8(self.pad2);
         buf.put_u32(self.event_type as u32);
