+++ src/simulation_management_with_reliability/remove_entity_reliable_pdu.rs
@@ -45,7 +45,7 @@ impl Pdu for RemoveEntityReliablePdu {
         buf.put_u8(self.required_reliability_service);
         buf.put_u16(self.pad1);
         buf.put_u8(self.pad2);
-        buf.put_u32(self.request_id as u32);
+        buf.put_u32(self.request_id);
     }
 
     fn deserialize(mut buffer: BytesMut) -> Result<Self, DISError>
