+++ src/simulation_management_with_reliability/start_resume_reliable_pdu.rs
@@ -52,7 +52,7 @@ impl Pdu for StartResumeReliablePdu {
         buf.put_u8(self.required_reliability_service);
         buf.put_u16(self.pad1);
         buf.put_u8(self.pad2);
-        buf.put_u32(self.request_id as u32);
+        buf.put_u32(self.request_id);
     }
 
     fn deserialize(mut buffer: BytesMut) -> Result<Self, DISError>
